requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==43.0.0
diff-match-patch==20230430
Django==5.0.6
django-allauth==0.63.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.13
django-colorfield==0.11.0
django-debug-toolbar==4.4.6
django-import-export==4.1.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-timezone-field==7.0
idna==3.7
kombu==5.3.7
ngrok==1.3.0
oauthlib==3.2.2
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13